Add files via upload
EMG_streaming_LUNA Is the specific MATLAB script that needs to be run to properly stream the data from the Noraxon software to LSL. The only thing that needs to be changed in this MATAB file is the file paths at the top of the file. These file paths just need to be changed to the correct file path in the computer running the program. The Noraxon software is also compatible with the AMTI force plate.